+++ code/files.js
@@ -562,9 +562,13 @@ function import_song(){
 		if(current_x_max>-999){
 			loading.xoffset = current_x_max + 4 - new_x_min;
 		}
-		/*if(songs.contains(loading.songname+"::notepad")){ //TODO - it should swap topbar for progress meter, clear the songlist and write out the notes in its place
-			post("\n\n\nSONG NOTES\n\n"+songs.get(loading.songname+"::notepad"));
-		}*/
+		if(songs.contains(loading.songname+"::notepad")){ //TODO - it should swap topbar for progress meter, clear the songlist and write out the notes in its place
+			sidebar.notification = songs.get(loading.songname+"::notepad");
+			set_sidebar_mode("notification");
+			// draw_sidebar();
+			// lcd_main.message("bang");
+			post("\n\n\nSONG NOTES\n\n"+ sidebar.notification+"\n\n");
+		}
 		loading.conncount = songs.getsize(loading.songname+"::connections");
 		loading.progress++;
 		loading.ready_for_next_action=loading.wait;//loading.bundling;