+++ code/mouse_helpers.js
@@ -1814,7 +1814,28 @@ function edit_typing(key){
 	redraw_flag.flag |= 2;
 }
 
+function edit_song_notes(key){
+	post("\ntyping,",key);
+	var caps = 0;
+	if(key==-2){
+		sidebar.text_being_edited = sidebar.text_being_edited+" ";
+	}else if(key==-4){
+		sidebar.text_being_edited = sidebar.text_being_edited+"£";
+	}else{
+		if(key>512) {
+			caps=1;
+			key-=512;
+			key-=32;
+		}
+		sidebar.text_being_edited = sidebar.text_being_edited + String.fromCharCode(key);
+	}
+	redraw_flag.flag |= 2;
+}
 
+function editted_song_notes(){
+	post("\ndone.",sidebar.text_being_edited);
+	set_sidebar_mode("none");
+}
 function static_mod_adjust(parameter,value){
 	//post("\nstatic mod adj",parameter[0],parameter[1],parameter[2],value,mouse_index);
 	//parameter holds paramno, blockno, voiceno